iOS support and tons of bug-fixes
