+++ TODO.md
@@ -1,8 +1,5 @@
 1. mobilesecurity_ print new nodeJS instructions
-2. File Manager - frida-fs download, BETA restore
-4. NOT working API: 18, 19, 25, 26 
-6. remove deepsource
-7. remove python stuff
-8. test iOS
-9. new readme file with nodejs instructions
-10. publish the npm package
+4. remove deepsource
+5. remove python stuff
+6. new readme file with nodejs instructions
+7. publish the npm package
