+++ views/templates/api_monitor.html
@@ -121,6 +121,9 @@
               <input type=submit 
                      value="Start APIs Monitor" 
                      class="btn btn-success">
+              <small class="form-text text-muted">
+              <font color=red><b>NOTE:</b></font> Some APIs Monitors (e.g. 18, 19, 25, 26) <b>may not work on some devices.</b> If the target app crashes, exclude them.
+              </small>
             </center>
           </form>
           <!-- Form end -->
